déplace du code pour améliorer la lisibilité
