+++ pkg/stats/statsActivitesParJour.go
@@ -1,0 +1,111 @@
+package stats
+
+import (
+	"context"
+	_ "embed"
+	"time"
+
+	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/pkg/errors"
+	"github.com/xuri/excelize/v2"
+)
+
+//go:embed resources/sql/select_activite_par_jour.sql
+var selectActiviteParJourSQL string
+
+type activiteParJour struct {
+	jour       time.Time
+	username   string
+	actions    int
+	recherches int
+	fiches     int
+	segment    string
+	err        error
+}
+
+func selectActiviteParJour(
+	ctx context.Context,
+	dbPool *pgxpool.Pool,
+	since time.Time,
+	to time.Time,
+	r chan activiteParJour,
+) {
+	defer close(r)
+	rows, err := dbPool.Query(ctx, selectActiviteParJourSQL, since.Truncate(day), to.Truncate(day))
+	if err != nil {
+		r <- activiteParJour{err: errors.Wrap(err, "erreur pendant la requête de sélection des activite/jour")}
+	}
+	for rows.Next() {
+		var activite activiteParJour
+		err := rows.Scan(
+			&activite.jour,
+			&activite.username,
+			&activite.actions,
+			&activite.recherches,
+			&activite.fiches,
+			&activite.segment,
+		)
+		if err != nil {
+			r <- activiteParJour{err: errors.Wrap(err, "erreur pendant la récupération des résultats/jour")}
+		}
+		r <- activite
+	}
+	if err := rows.Err(); err != nil {
+		r <- activiteParJour{err: errors.Wrap(err, "erreur après la récupération des résultats/jour")}
+	}
+}
+
+func writeActiviteJourToExcel(f *excelize.File, sheetName string, ligne activiteParJour, row int) error {
+	var i = 1
+	err := writeString(f, sheetName, ligne.jour.Format(time.DateOnly), i, row)
+	i++
+	if err != nil {
+		return errors.Wrap(err, "erreur pendant l'écriture du jour")
+	}
+	err = writeString(f, sheetName, ligne.username, i, row)
+	if err != nil {
+		return errors.Wrap(err, "erreur pendant l'écriture du nom de l'utilisateur")
+	}
+	i++
+	err = writeInt(f, sheetName, ligne.actions, i, row)
+	if err != nil {
+		return errors.Wrap(err, "erreur pendant l'écriture du nombre d'actions")
+	}
+	i++
+	err = writeInt(f, sheetName, ligne.recherches, i, row)
+	if err != nil {
+		return errors.Wrap(err, "erreur pendant l'écriture du nombre de recherches")
+	}
+	i++
+	err = writeInt(f, sheetName, ligne.fiches, i, row)
+	if err != nil {
+		return errors.Wrap(err, "erreur pendant l'écriture du nombre de fiches")
+	}
+	i++
+	err = writeString(f, sheetName, ligne.segment, i, row)
+	if err != nil {
+		return errors.Wrap(err, "erreur pendant l'écriture du nombre de segments")
+	}
+	return nil
+}
+
+func writeActivitesJoursToExcel(xls *excelize.File, pageIndex int, activites chan activiteParJour) error {
+	var sheetName, err = createSheet(xls, "Activité par jour", pageIndex)
+	if err != nil {
+		return errors.Wrap(err, "erreur lors de la création de la feuille excel")
+	}
+	var row = 1
+	if activites != nil {
+		for ligne := range activites {
+			if ligne.err != nil {
+				return ligne.err
+			}
+			err := writeActiviteJourToExcel(xls, sheetName, ligne, row)
+			if err != nil {
+				return err
+			}
+			row++
+		}
+	}
+	return nil
+}
